enroll: use realm form the HOST$ entry in the keytab
Currently the realm from the first entry in the keytab was used. This
might fail if the keytab has entries from other realms as well. With
this patch adcli will first try to find a HOST$ entry in the keytab and
take its realm because this will most probably be the principal used to
authenticate against the AD DC of realm the client is joined to. If no
realm can be found this way the realm of the first entry is still used
as a fallback.