enroll: restore SELinux file context of keytab files
If SELinux is enable in a system the file context of the keytab file
might be important because it will restrict access for confined services
and applications.

Typically the system keytab /etc/krb5.keytab is handled by the system
policy. But adcli will create e.g. /etc/krb5.keytab.domain.name for MSAs
which should get the keytab specific file context as well. This patch
adds a SELinux policy module for adcli which set the file context for
files matching the pattern /etc/krb5.keytab.* and
/etc/sssd/krb5.keytab.* .  Unfortunately wildcards are not supported for
automatic transitions in the kernel so that adcli has to call
selinux_restorecon() manually on new or changed keytab files.

Resolves: https://issues.redhat.com/browse/RHEL-78631